refact: 리워드 공통 로직 Common으로 분리
#17
