+++ Application/Reward/src/main/java/com/example/Reward/Receipt/Controller/ReceiptController.java
@@ -1,5 +1,6 @@
 package com.example.Reward.Receipt.Controller;
 
+import com.example.Reward.Common.Service.GiveStockService;
 import com.example.Reward.Receipt.Dto.in.RewardRequestDTO;
 import com.example.Reward.Receipt.Dto.out.AnalyzeReceiptDTO;
 import com.example.Reward.Receipt.Dto.out.CheckReceiptResponseDTO;
@@ -20,9 +21,11 @@ import java.math.BigDecimal;
 @Tag(name="영수증 API")
 public class ReceiptController {
     private final ReceiptService receiptService;
+    private final GiveStockService giveStockService;
 
-    public ReceiptController(ReceiptService receiptService) {
+    public ReceiptController(ReceiptService receiptService, GiveStockService giveStockService) {
         this.receiptService = receiptService;
+        this.giveStockService = giveStockService;
     }
 
     @GetMapping("/enterprise")
@@ -56,8 +59,8 @@ public class ReceiptController {
     @PostMapping("/")
     @Operation(description = "사용자 확인 후 영수증 정보 저장 및 리워드 제공")
     public ResponseEntity<RewardResponseDTO> rewardStock(@RequestHeader("memberId") String memberId, @RequestBody RewardRequestDTO rewardRequestDTO) {
-        Integer priceOfStock = receiptService.getPrice(rewardRequestDTO.getEnterpriseName());
-        Double amountOfStock = receiptService.calDecimalStock(priceOfStock);
+        Integer priceOfStock = giveStockService.getPrice(rewardRequestDTO.getEnterpriseName());
+        Double amountOfStock = giveStockService.calDecimalStock(priceOfStock);
         RewardResponseDTO rewardResponseDTO = receiptService.giveStockAndSaveReceipt(Long.valueOf(memberId), rewardRequestDTO, priceOfStock, amountOfStock);
         return ResponseEntity.ok(rewardResponseDTO);
     }
