+++ Application/Reward/src/main/java/com/example/Reward/Receipt/Service/ReceiptService.java
@@ -6,6 +6,7 @@ import com.example.Reward.Advertisement.Webclient.GeneratedToken;
 import com.example.Reward.Common.Entity.Event;
 import com.example.Reward.Common.Kafka.GiveStockProduceDto;
 import com.example.Reward.Common.Repository.EventRepository;
+import com.example.Reward.Common.Service.GiveStockService;
 import com.example.Reward.Receipt.Dto.in.RewardRequestDTO;
 import com.example.Reward.Receipt.Dto.out.*;
 import com.example.Reward.Receipt.Dto.webClient.PresentPriceDTO;
@@ -30,23 +31,16 @@ import java.util.*;
 @Service
 @RequiredArgsConstructor
 public class ReceiptService {
-
+    private final GiveStockService giveStockService;
     private final EventRepository eventRepository;
     private final ReceiptLogRepository receiptLogRepository;
     private final AmazonS3 amazonS3;
+    private final WebClient webClient;
     @Value("${cloud.aws.s3.bucket}")
     private String bucket;
-    private final WebClient webClient;
     @Value("${x-ocr-secret}")
     private String ocrSecret;
     private static final String OCR_BASE_URL = "https://1l8mnx9ap5.apigw.ntruss.com";
-    private static final String STOCK_BASE_URL = "https://openapi.koreainvestment.com:9443";
-    @Value("${app.key}")
-    private String appKey;
-    @Value("${app.secretkey}")
-    private String appSecret;
-    private final KafkaTemplate<String,Object> kafkaTemplate;
-    private final GeneratedToken generatedToken;
     private final GetLongestCommonSubstring getLongestCommonSubstring;
 
     public GetEnterpriseListDTO getEnterpriseList() {
@@ -156,30 +150,9 @@ public class ReceiptService {
         return "";
     }
 
-
-    public Integer getPrice(String enterpriseName) {
-        String stockCode = eventRepository.findByEnterpriseNameContainingAndContentId(enterpriseName, 2L).getStockCode();
-        System.out.println(stockCode);
-        Mono<PresentPriceDTO> response = webClient.get()
-                .uri(STOCK_BASE_URL + "/uapi/domestic-stock/v1/quotations/inquire-price?FID_COND_MRKT_DIV_CODE=J&FID_INPUT_ISCD={param}", stockCode)
-                .header("authorization", "Bearer " + generatedToken.getAccessToken())
-                .header("appkey", appKey)
-                .header("appsecret", appSecret)
-                .header("tr_id", "FHKST01010100")
-                .retrieve()
-                .bodyToMono(PresentPriceDTO.class);
-        PresentPriceDTO result = response.block();
-        return result.getOutput().getStck_prpr();
-    }
-
-    public Double calDecimalStock(Integer priceOfStock) {
-        BigDecimal price = new BigDecimal(Integer.toString(priceOfStock));
-        return BigDecimal.valueOf(100).divide(price, 6, BigDecimal.ROUND_DOWN).doubleValue();
-    }
-
     @Transactional
     public RewardResponseDTO giveStockAndSaveReceipt(Long memberId, RewardRequestDTO rewardRequestDTO, Integer priceOfStock, Double amountOfStock) {
-        giveStock(memberId, rewardRequestDTO, priceOfStock, amountOfStock);
+        giveStockService.giveStock(memberId, rewardRequestDTO, priceOfStock, amountOfStock);
         Event event = eventRepository.findByEnterpriseNameContainingAndContentId(rewardRequestDTO.getEnterpriseName(), 2L);
         event.setRewardAmount(event.getRewardAmount()-amountOfStock);
         eventRepository.save(event);
@@ -202,18 +175,5 @@ public class ReceiptService {
                 .amount(amountOfStock)
                 .build();
         return rewardResponseDTO;
-    }
-
-    public void giveStock(Long memberId, RewardRequestDTO rewardRequestDTO, Integer priceOfStock, Double amountOfStock) {
-        String stockCode = eventRepository.findByEnterpriseNameContainingAndContentId(rewardRequestDTO.getEnterpriseName(), 2L).getStockCode();
-        GiveStockProduceDto giveStockProduceDTO = GiveStockProduceDto.builder()
-                .memId(memberId)
-                .enterpriseName(rewardRequestDTO.getEnterpriseName())
-                .code(stockCode)
-                .price(priceOfStock)
-                .amount(amountOfStock)
-                .build();
-
-        kafkaTemplate.send("test-mo", giveStockProduceDTO);
     }
 }
